+++ packages/build-utils/src/types.ts
@@ -196,6 +196,12 @@ export interface StartDevServerSuccess {
    * shut down the dev server once an HTTP request has been fulfilled.
    */
   pid: number;
+
+  /**
+   * An optional function to shut down the dev server. If not provided, the
+   * dev server will forcefully be killed.
+   */
+  shutdown?: () => Promise<void>;
 }
 
 /**
@@ -434,13 +440,6 @@ export interface Cron {
   schedule: string;
 }
 
-// TODO: Proper description once complete
-export interface Flag {
-  key: string;
-  defaultValue?: unknown;
-  metadata: Record<string, unknown>;
-}
-
 /** The framework which created the function */
 export interface FunctionFramework {
   slug: string;
@@ -465,7 +464,7 @@ export interface BuildResultV2Typical {
   framework?: {
     version: string;
   };
-  flags?: Flag[];
+  flags?: { definitions: FlagDefinitions };
 }
 
 export type BuildResultV2 = BuildResultV2Typical | BuildResultBuildOutput;
@@ -485,3 +484,30 @@ export type ShouldServe = (
 export type StartDevServer = (
   options: StartDevServerOptions
 ) => Promise<StartDevServerResult>;
+
+/**
+ * TODO: The following types will eventually be exported by a more
+ *       relevant package.
+ */
+type FlagJSONArray = ReadonlyArray<FlagJSONValue>;
+
+type FlagJSONValue =
+  | string
+  | boolean
+  | number
+  | null
+  | FlagJSONArray
+  | { [key: string]: FlagJSONValue };
+
+type FlagOption = {
+  value: FlagJSONValue;
+  label?: string;
+};
+
+export interface FlagDefinition {
+  options?: FlagOption[];
+  origin?: string;
+  description?: string;
+}
+
+export type FlagDefinitions = Record<string, FlagDefinition>;