+++ packages/cli/src/util/dev/builder.ts
@@ -361,11 +361,13 @@ export async function executeBuild(
           await oldAsset.fn.destroy();
         }
 
+        const ZipFile = asset.zipBuffer || (await asset.createZip());
+
         asset.fn = await createFunction({
-          Code: { ZipFile: asset.zipBuffer },
+          Code: { ZipFile },
           Handler: asset.handler,
           Runtime: asset.runtime,
-          MemorySize: asset.memory || 3008,
+          MemorySize: asset.memory || 3009,
           Environment: {
             Variables: {
               ...vercelConfig.env,