+++ packages/cli/src/util/link/setup-and-link.ts
@@ -198,6 +198,7 @@ export default async function setupAndLink(
         projectSettings: {
           ...localConfigurationOverrides,
           sourceFilesOutsideRootDirectory,
+          rootDirectory,
         },
         autoAssignCustomDomains: true,
       };