+++ packages/remix/test/fixtures-vite/02-interactive-remix-routing-v2/app/data/routes/basicRoutes/index.tsx
@@ -1,0 +1,3 @@
+import basicRoutes from './basicRoutes';
+
+export default basicRoutes;